Removed unneccessary file.
Added Qt Creator project file for debugging, addressed some width-conversion related warnings.

Fixed warning related to converting wide data type to a narrow data type.

Fixed warning related to converting wide data type to a narrow data type.

Fixed warning related to converting wide data type to a narrow data type.

Fixed warning related to converting wide data type to a narrow data type.

Added HAVE_CONFIG_H definition to CMakeLists.txt instead of relying on it being defined in the source.

Removed duplicate definition of HAVE_CONFIG_H

Fixed warning related to converting wide data type to a narrow data type.

Fixed warning related to converting wide data type to a narrow data type.

Fixed warning related to converting wide data type to a narrow data type.

Fixed warning related to converting wide data type to a narrow data type.

Fixed warning related to converting wide data type to a narrow data type.

+++ libdap2/dapcvt.c
@@ -195,8 +195,8 @@ dapcvtattrval3(nc_type etype, void* dst, NClist* src)
 {
     int i,ok;
     NCerror  ncstat = NC_NOERR;
-    unsigned int memsize = nctypesizeof(etype);
-    unsigned int nvalues = nclistlength(src);
+    size_t memsize = nctypesizeof(etype);
+    size_t nvalues = nclistlength(src);
     char* dstmem = (char*)dst;
 
     for(i=0;i<nvalues;i++) {
