+++ ncgen/data.h
@@ -36,11 +36,11 @@ typedef union Constvalue {
     float floatv;               /* NC_FLOAT*/
     double doublev;             /* NC_DOUBLE*/
     struct Stringv {		/* NC_STRING*/
-	int len;
+    size_t len;
 	char* stringv; 
     } stringv;
     struct Opaquev {     /* NC_OPAQUE*/
-	int len; /* length as originally written (rounded to even number)*/
+    int len; /* length as originally written (rounded to even number)*/
 	char* stringv; /*as  constant was written*/
 		      /* (padded to even # chars >= 16)*/
 		      /* without leading 0x*/
